+++ Modules/ChattingModule/Database/Migrations/2022_08_21_033729_add_col_to_channel_user_table.php
@@ -22,6 +22,10 @@ class AddColToChannelUserTable extends Migration
             });
         }
 
+        if (! Schema::hasTable('channel_lists')) {
+            return;
+        }
+        
         Schema::table('channel_lists', function (Blueprint $table) {
             $table->dropColumn('is_read');
         });
@@ -34,6 +38,10 @@ class AddColToChannelUserTable extends Migration
      */
     public function down()
     {
+        if (! Schema::hasTable('channel_users')) {
+            return;
+        }
+        
         Schema::table('channel_users', function (Blueprint $table) {
 
         });
