+++ Modules/FieldItems/Database/Migrations/2023_03_15_062023_add_company_id_to_item_files_table.php
@@ -28,6 +28,10 @@ return new class extends Migration
      */
     public function down()
     {
+        if (! Schema::hasTable('item_files')) {
+            return;
+        }
+        
         Schema::table('item_files', function (Blueprint $table) {
                 $table->dropForeign(['company_id']);
                 $table->dropColumn('company_id');
