+++ Modules/Purchase/Database/Migrations/2024_05_09_094658_add_purchase_terms_to_purchase_setting_table.php
@@ -26,6 +26,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (! Schema::hasTable('purchase_setting')) {
+            return;
+        }
+        
         Schema::table('purchase_setting', function (Blueprint $table) {
             $table->dropColumn('purchase_terms');
         });
